Update to Imaging Release API
